Reuse `percentageDistribution()`
