+++ frontend/src/scenes/experiments/utils.ts
@@ -36,6 +36,17 @@ export function formatUnitByQuantity(value: number, unit: string): string {
     return value === 1 ? unit : unit + 's'
 }
 
+export function percentageDistribution(variantCount: number): number[] {
+    const basePercentage = Math.floor(100 / variantCount)
+    const percentages = new Array(variantCount).fill(basePercentage)
+    let remaining = 100 - basePercentage * variantCount
+    for (let i = 0; remaining > 0; i++, remaining--) {
+        // try to equally distribute `remaining` across variants
+        percentages[i] += 1
+    }
+    return percentages
+}
+
 export function getMinimumDetectableEffect(
     metricType: InsightType,
     conversionMetrics: FunnelTimeConversionMetrics,
